refactor: simplify CLI entry point
