+++ cmd/args.go
@@ -1,0 +1,68 @@
+package cmd
+
+import (
+	"path"
+	"strings"
+)
+
+func mapAliasArgs(argv []string) []string {
+	if len(argv) == 0 {
+		return nil
+	}
+
+	args := argv[1:]
+	if !isSwxAlias(argv[0]) {
+		return args
+	}
+
+	if len(args) > 0 {
+		// Let users generate completion scripts for `swx` itself.
+		if args[0] == "completion" {
+			return args
+		}
+
+		// Cobra shell completion calls these internal commands.
+		// Prefixing `project console` preserves swx-as-console behavior for completions.
+		if args[0] == "__complete" || args[0] == "__completeNoDesc" {
+			aliasedCompletionArgs := make([]string, 0, len(args)+2)
+			aliasedCompletionArgs = append(aliasedCompletionArgs, args[0], "project", "console")
+			aliasedCompletionArgs = append(aliasedCompletionArgs, args[1:]...)
+
+			return aliasedCompletionArgs
+		}
+	}
+
+	// When invoked via the `swx` symlink, forward everything to `project console`.
+	aliasedArgs := make([]string, 0, len(args)+3)
+	aliasedArgs = append(aliasedArgs, "project", "console")
+
+	if len(args) == 0 {
+		aliasedArgs = append(aliasedArgs, "list")
+	} else {
+		aliasedArgs = append(aliasedArgs, args...)
+	}
+
+	return aliasedArgs
+}
+
+func isSwxAlias(binaryPath string) bool {
+	return strings.EqualFold(commandNameFromBinaryPath(binaryPath), "swx")
+}
+
+func commandNameFromArgs(argv []string) string {
+	if len(argv) == 0 {
+		return rootCmd.Use
+	}
+
+	return commandNameFromBinaryPath(argv[0])
+}
+
+func commandNameFromBinaryPath(binaryPath string) string {
+	normalizedPath := strings.ReplaceAll(binaryPath, "\\", "/")
+	binaryName := strings.TrimSuffix(path.Base(normalizedPath), path.Ext(normalizedPath))
+	if binaryName == "" {
+		return rootCmd.Use
+	}
+
+	return binaryName
+}
