+++ main.go
@@ -2,10 +2,11 @@ package main
 
 import (
 	"context"
+	"os"
 
 	"github.com/shopware/shopware-cli/cmd"
 )
 
 func main() {
-	cmd.Execute(context.Background())
+	os.Exit(cmd.Execute(context.Background()))
 }
